Fix lint and typos

+++ src/main/scala/gitbucket/core/controller/ApiController.scala
@@ -381,7 +381,7 @@ trait ApiControllerBase extends ControllerBase {
   get("/api/v3/repos/:owner/:repository/issues/:id/comments")(referrersOnly { repository =>
     (for{
       issueId  <- params("id").toIntOpt
-      comments =  getCommentsForApi(repository.owner, repository.name, issueId.toInt)
+      comments =  getCommentsForApi(repository.owner, repository.name, issueId)
     } yield {
       JsonFormat(comments.map{ case (issueComment, user, issue) => ApiComment(issueComment, RepositoryName(repository), issueId, ApiUser(user), issue.isPullRequest) })
     }) getOrElse NotFound()
